move plot sample convenience function to visualization module
